Refine entry publication quota
The previous approach prevented published entries from being updated
when the published entries quota was exhausted. Instead we need to
check whether the quota would be exceeded under the assumption that
the current entry was published. As a side effect, published entries
cannot be updated if the quota already is exceeded, which is the
desired behaviour.

To come up with a sensible ui, the editor now requires information
about both the current and the assumed state of the quota. Therefore,
the generic `editor/quotas/:name endpoint` has been replaced by a
dedicated `entry_publications` endpoint which supports both checking
whether the entry can be published and performing the publication.

* add EntryPublication poro to encapsulate publication logic
* add EntryPublicationsController to check and publish entries
* add EntryPublication Backbone model
* turn generic QuotaDecoratorView into
  EntryPublicationQuotaDecoratorView
* make PublishEntryView use EntryPublication Backbone model
* remove Editor::QuotasController
* remove RevisionsController#create which is now handled via
  EntryPublications#create
* add verify_not_exceeded! method to Quota class