Simulate Media Queries in Editor
When resizing the editor sidebar, the entry can be displayed with
mobile width. Still media queries do not apply causing breaking
layouts. Intead, introduce an alternative application.css file for the
editor which uses implentations of the break point mixins that are
controlled by css classes. Then in the EntryPreviewView update these
classes on resize based on the width of the preview - not the browser
window.