Remove window onload timeout
pull request #119 introduced a timeout which resolves the ready
promise if window.onload is not fired after 10 seconds, to solve #118
in Safari.  This change introduced a new problem though: When the dom
ready event takes longer than 10 seconds we start while some of the
widgets are not yet defined. This happens mostely in
development. Still, since #131 solved the onload issue in Safari, we
revert the timeout hack.