Prevent incorrectly called page type hook on manual start
If manual start is enabled, the activated hook of the first page is
not supposed to be triggered if another page has already been
navigated to via the url fragment.

The page widget listens for its `deactivate` event to cancel delayed
page type hooks. The event name was wrong though, causing the page
hook to fire anyway.