Account delete should delete themings/folders
Account destroy did not get rid of the associated theming. There is no
dependent: :destroy on the relation.

This is problematic because in the Edit Entry screen, a list of themings
is needed for the theming dropdown. When a theming does not have a
corresponding account, line 23 raises undefined method 'name' for
nil:NilClass.

The same holds for an account's folders.

fixes #476