introduce .logo_header
In the current theme, the logo header on the first page is styled using
a CSS pseudo element. This can be troublesome to work with, since the
element does not have a class.

It is also troublesome to discover that this element exists and change
it according to one's own needs. The CSS-rule is quite exact and does
not allow for much flexibility.

This patch aims to change that.

By introducing a new div with class name "logo_header" it becomes much
easier to style this element using CSS, and to override it in a custom
theme. The element becomes easier to inspect in the browser and change
some style attributes there. Lastly it's evident what this element does,
and ties in directly with the image name that's already in use.

It does mean that there is now a DIV in each page. This is an acceptable
tradeoff to me; in fact I prefer this since it will allow me to style
the header on the last page too. It gives me more power and flexibility.

@tf thoughts?