Improve setting of near_bottom class in header
When scrolling back to a page the class may not be set since we are at
the bottom of the page. Rely completely on the scroller events of the
current page.

Detect current page via comparison with `slides.currentPage`, since
the `active` class on page is set with a delay.