Add generator specs
Using `ammeter`, this pull request adds test for generators.

We are testing output of the generators, but up to a point. We're not
recreating the full ActiveAdmin initializer, for example.

This pull request also marks the first time we're not relying on the
magical inclusion of support files from `spec/support`. This will boost
performance of all tests since we're only requiring the files we
actually need. In future patches we should move towards this structure
for all tests, and stop automatically requiring all files under support.

The `type: :generator` will automatically load the shared context
defined in `spec/support/helpers/shared_contexts.rb`.