Generate seed data for files
Add a common json partial that is used by both the `FilesHelper`,
which renders seed data for the editor, and the `EntryJsonSeedHelper`,
which renders seed data for published entries.

Let file types specify a custom partial that is included in these
representations.

Add id, parent file and basename to file seed data.