Only render persisted pages in editor preview
This is important for the Redux integration since unpersisted pages
are not yet represented in the store. We need to make sure store data
is present once a React based page types is initialized.

Add a timeout when adding pages to the subset collection, to ensure
that the event handlers that synchronize Backbone collections with the
Redux store have completed their work.