change Expires to Cache-Control header on S3
Backport of #753

Expires is HTTP/1.0
Cache-Control is HTTP/1.1 and supersedes Expires.
Cache-Control offers everything from Expires and a lot more.
Support for it is actually better than for Expires.

In our case, it's important to add the `public` to the header as well, so that the files will
be cached by proxies.