Fix ensureValueDefined option of SelectInputView
The change in #725 made sure that the first option is always displayed
as a default value in a select box without blank item. This broke the
`ensureValueDefined' option: A value would only be defined if the
select box did not have selected item.

Instead explicitly check that if the model has a value defined for the
property. If not, calling `save` is enough now since the first item
has already been selected.