Allow specifying min width of title in split layout
Add a modifier class `title-for_split_layout` that ensures that the
title is contained inside the left column of a split layout. Add a css
variable to define the min width the title is supposed to have in
split layout.

The two page types that use split layout, chart and embedded video,
both measure the space on the right side of the title to toggle split
layout. So when the break point is reached and the title width
shrinks, they will position the embed inside the right column.