Fix page title hysteresis
On choosing a page link, properly update title when it should be an
empty string. Without this, e.g. in ReferenceInputView, `title` will
not update in certain situations, because page#title() returns
`'undefined'` due to `''` falsiness in Javascript.

To reproduce:
```
this.configuration.get('title') === ''
this.configuration.get('additional_title') === undefined
```