Paperclip :pageflow_filesystem_root interpolation
Previously, this interpolation would only be defined in development, test and production environments. A staging environment did not have this interpolation which would lead to an error. Changed the existing code to a switch inside the definition, so that an interpolation value always exists.

The special cases for Rails.env.test should probably move into a support file under spec.