Use page from URL hash es landing page
Previously when opening a URL with a page id fragment, the entry's
first page would shortly be visible followed by an instant transition
to the target page.

This is especially problematic for analytics purposes as the short
visit to the first page is counted as a normal page impression.

This changes ensures the page given by the id in the url fragment is
visited right away.