Improve handling of externally controlled player
When the media player is paused by an external event (i.e. a different
player library that has direct access to the media element), we want
to make sure that the `PAUSED` action does not lead to another call of
`pause` on the media element.

We can distinguish this from the case where `PAUSE` caused
`shouldPlay` to become `false` by checking whether `isPlaying` is
already false and skip calling `pause` on the media player.

By mistake, `isPlaying` was also set to `false` on
`FADE_OUT_AND_PAUSE` before. This needs to be removed since otherwise
now the `pause` would be skipped

REDMINE-16148, REDMINE-14305