Allow registering scroller events on wrapper element
When clipping the scroller at the bottom we still want to be able to
interact with the scroller on the whole page. Wrap the scroller into
an element that remains at full height even if a bottom margin is
applied to the scroller and make iScroll register its mouse and
pointer events on that wrapper.

As a result the page can still be scrollen when performing pointer
down in the info box area of the waveform player controls, even
though the scroller content is clipped above the waveform.