replace simple jQuery with native JavaScript
element.classList is quite widely supported:
http://caniuse.com/#search=classlist

Legacy IE does not support multiple parameters for the add() and
remove() methods. I could solve this by using two lines for the removal.

I know there is a solid dependence on jQuery but for these simple class
manipulations we shouldn't need to load all of jQuery?