read the attribute instead of super
super is a bit counter-intuitive since the method might not exist on the superclass, but we're just relying on the message being picked up anyway. By reading the attribute the intent is made more clear.

We are not using `read_attribute` but the brackets, because the latter raises an exception when the attribute cannot be found. The rails style guide recommends so. https://github.com/bbatsov/rails-style-guide#read-attribute