Handle async tech loading in VideoJS 6
VideoJS 6 initialized the HTML5 tech asynchronously. Interaction with
the API is only possible after the `ready` callback has been invoked.

When updating video related attributes in the editor, attempts to
interact with the player shortly after its recreation lead to
exceptions which prevented changes to be saved to the server.