Catch promise if play was aborted
Chrome returns a promise from `play` that is rejected if another
operation (like calling `pause` or updating the source) happens before
playing started. If the promise rejection is not handled, the
operation that caused `play` to abort will fail with an
exception. Code following the operation will not be executed. Catch
and ignore the promise to prevent this.

Source updates related to DASH appear to trigger this case.