Fix jQuery error in node_package specs on Travis
Calling jQuery's `append` method started to cause an error of the
form

    'undefined' is not a function (evaluating 'elem.getAttribute( "type" )')

in a jQuery function called `disableScript` which tries to change the
`type` attribute of script tags while parsing HTML. I could not
reproduce the error locally. Also it is not clear to me why there
should be `script` elements inside the newly created div element.

As a workaround we can use DOM API instead of jQuery to setup the
fixture page element.