Fix page transition effects in Chrome 66
Ensure `content` and `backgroundArea` only have transitions applied
when they are actually being animated. Otherwise they might be visible
when we want to move them out of the view quickly.

For page transitions CSS this means only applying `transition` when

- the `active` class is present (for animating a page into view)

- an `animate-out-*` class is present (for animating a page out of
  view when it no longer has the `active` class.

`animate-in-*` classes are applied before the `active` class and are
used to move the page out of the viewport before the animation
begins. This needs to happen immediately.

For historic reasons, a redundant transition declaration was applied
in `slideshow.scss` that had to be removed. The different page
transitions need to handle this themselves.

Also the slim player controls applied a transition to the page
background to fade the pseudo element that is used as a shadow behind
the player controls. It is enough to apply the transition to the
pseudo element itself.