Consolidate migrations
While it's ok for apps to maintain a full migration history, there is
no need for new apps to replay everything.

Replace all existing migrations with one that creates the final schema
and simply does nothing if `pageflow_` tables already exist.

Apps need to be updated to Pageflow 12.2 before updating to 13.0.