Ensure mobile video poster is displayed on Android
Before the poster was only displayed if the `native_video_player`
feature was present. This is only the case on iPhone.

Now we show the poster on all phone platforms, but hide it while
videos are playing if the `native_video_player` feature is not
present. Otherwise the video being played inline would not be visible.