Improve initial password workflow for invitations
When a user clicks the link in an invitation mail, we used to send
them to the edit password page. The wording of that page is tailored
to the case when users actually want to recover their password - not
for setting a password while accepting an invitation.

In particular the view contained a "Sign in" link which lead to the
normal sign in page. Invited users sometimes clicked the link instead
of saving the password.

Subclass `Devise::PasswordsController` to use a customized view. Adapt
the wording. Remove the "Sign in" link.