Make maximum character count for text input fields configurable
This sets a default maxLength option for text input fields with a value of 255 but also accepts a different value passed in via the options hash.
Validation only takes effect for new and maxLength-conform data.
Legacy data which exceeds the specified limit will remain unvalidated but new data must conform to the specified maxLength option.