Use polyfill to fix Array#reverse in Safari 12
In Safari 12, creating an array via a literal and calling
`Array#reverse` on it causes arrays created from an identical literal
to be reversed as well from that point on. An internal shared
representation of the array literal gets mutated.

https://stackoverflow.com/questions/52390368/array-state-will-be-cached-in-ios-12-safari-is-it-a-bug-or-feature/52392901#52392901

This polyfill detects the bug and overrides `Array#reverse` with a
fixed implementation.

The bug caused video text tracks to switch positions from bottom to
top on each update.