Revert attachment name change
In order to keep compatibility with S3 for existing attachment files we need to keep the attachment name for hosted files.
For image files (where the buckets are named "processed_attachments") this introduces a new interpolation which can be
defined on the model. In our case it's only defined on the ImageFile-model.

The self.columns class method on HostedFile is also being removed and the documentation updated accordingly.