Trigger resize hooks on cookie notice dismiss
Backport of #1101

Previously, dismissing the cookie notice removed the div but did not resize the banner image,
which left an empty div at the bottom of the page.
Explicitly calling the resize hook in pageflow.widgets.use takes care of resizing the banner image
back to the full available height.