Ensure same file can be selected multiple times
jQuery file uploader needs the file input's `change` event to fire to
start its process. `replaceFileInput` needs to be left at it's truthy
default value to ensure this happens when selecting the same file
multiple times in a row.

Then we can no longer use `ui.input` to access the input, though,
since that is only set once when `bindUIElements` is called. After the
uploader replaces the input, the reference becomes stale. Instead use
`Backbone.View#$` to look up the input each time and revert to
initializing the file uploader on the form element (`this.$el`)
instead of the file input.

Remove unused option `autoUpload`. The uploader source does not
mention this option. I think it is part of the UI component which
comes with the uploader, but is not used by Pageflow.