Create new entries with default locale
Customers who have set their language to english have reported that
the default language for new stories is still German.
This behavior was perceived as a bug.
This PR introduces a new column default_locale on account/theming level
where a default language for new stories can be specified (defaults to German),
which is then used for new stories.

For new accounts the default_locale will initially get set to the locale the user
specified during signup.