Do not use asset_host in entry stylesheet link tags
Backport of #1107

`Configuration#public_entry_request_scope` and
`Configuration#public_entry_redirect` can be used to restrict which
entries are accessible under which hostname. This also applies to
entry stylesheet requests. Therefore, when an asset host is
configured, we need to make sure that entry stylesheet requests are
not blocked.

`stylesheet_link_tag` always uses the asset host when a relative URL
is passed. We could pass an absolute URL using the hostname of the
current request, but that might cause problems with caching.

Enforce always using a relative URL by generating the link tag
manually.

REDMINE-16292