Ensure entries can be filtered by publication date
Trying to filter entries by publication date in the entry admin index
list caused an error of the form:

    Column 'published_at' in on clause is ambiguous

The combination of the `includes(:published_revision)` in
`scoped_collection` and the `:published_revision_published_at` filter
caused the revisions table to be joined twice, making the unqualified
column `published_at` in the query of the `Revision.published` scope
ambiguous.

Since there is no documented way to obtain the current table alias,
the `published` query has been rewritten to use hash syntax instead of
an SQL fragment.

The comparison of `published_until` can be replaced with a date
range. Since `published_at` is only ever set to the current time, its
value can never be in the future. Checking for its presence is
therefore equivalent to checking whether it lies in the past.

Duplicating the `published_at IS NOT NULL` check, to make precendence
of `AND` and `OR` operators explicit.

REDMINE-16490