Prevent preload request from detached media element
A detached DOM element is used to build the HTML for
`MediaTag`. Setting `preload` to `auto` in the detached element,
triggers a separate request in Firefox 67 which afterwards keeps a
network connection open.

Set `preload` to `none` and perform text replacement in the string
which is passed to `dangerouslySetInnerHTML`.

REDMINE-16834