Add text position center option
Page types can now opt in to a new text position variant called
"center" by passing the `supportsTextPositionCenter` option to the
`general` configuration editor inputs group and setting the
`supports_text_position_center` class on their wrapper element.

Text stays left aligned but is moved to the center of the page. The
shadow turns into a transparent single-color overlay.

So far only the built in page types support this.

Adapt select input view to display the default option if the current
value is not among the options. This ensures that the drop down
displays "Left" even if "Center" has been selected before changing the
page type.

REDMINE-16909