Remove character set from setup schema migration
While it is debatable whether it is good to have a default charset in
pageflow's migrations, it is quite clear that utf8, which is an alias
for utf8mb3, isn't a good default: It doesn't encode emoji, and
emoji have become popular.

For this reason, we drop utf8 as default charset.

REDMINE-17010