Let entry types provide frontend apps to render entries
Add concern which gives controllers the possibility to delegate to a
Rack app.

This mechanism does not work, though, in controller specs (which are
based on `ActionController::TestCase`). We therefore have to extract
the corresponding specs into request specs (which are based on
`ActionDispatch::IntegrationTest`).

Move paged specific functionality into a temporary
`Pageflow::PagedEntriesController`, before extracting it the entry
type engine in a future commit.

REDMINE-17208