Add needed command to node package development docs
When linking a local pageflow gem from a host application (for the
first time?), we need to run `yarn install` in the host application to
pick up the linked packages in a way that doesn't let the app crash.

Also, this replaces two 'or' by 'and', because 'or' confused me as to
which order would be valid.

REDMINE-17257