Let entry types define entry Backbone model
`Entry` Backbone model currently contains a lot of entry type specific
code, which needs to be extracted. Letting entry types extend the
model, will an easy way to do this without changing interfaces in the
paged editor.

Add `setupFromEntryTypeSeed` method to entry model, which can be used
to read the entry type specific seed. For now we only use it to store
the seed data rendered by the editor seed fragment, such that the
entry preview view can read it.

REDMINE-17260