Consider entry_type in FeatureTarget#enabled_feature_names
To allow registering features inside entry type blocks, when building
a configuration object for an entry, we need to make sure that we do
not try to enable features which are not registered in the entry's
configuration object. We therefore need to make sure, that features
which are enabled by default do not show up in an entry's
`enabled_feature_names` if the feature has been registered for a
different entry type.

So far, `enabled_feature_names` relied on the global configuration to
get the list of all feature names. We need to use an entry specific
configuration.

`enabled_feature_names` is called in two different scenarios:

* From seed helpers which pass a list of enabled feature names. Here
  we can simply call `config_for` to construct the entry specific
  configuration.

* From `config_for` to find out which feature blocks to evaluate.  To
  prevent an infinite loop, we make `config_for` pass in the current
  configuration object so that `enabled_feature_names` can read the
  list of feature names directly from there.

REDMINE-17275