Rename files_json_seeds helper to editor_files_json_seed
We want to extract a files json seed helper from
`pageflow/entry_json_seed/_entry.json.jbuiler` to reuse it for
scrolled entries.

Since we already started to move editor specific helpers inside the
`Pageflow::Editor` namespace, it makes sense to move the helper which
is used for editor seed data to a new `Editor::FilesHelper`. That way
the name `files_json_seed` becomes available again.

At the moment, both "seeds" and "seed" is used in helper names. The
singular version seems to be used a bit more frequent in newer
code. So we continue to move things in this direction.

REDMINE-17301