Fix section index and references to adjacent chapters
In #1298, the `Chapter` component was extracted from the `Entry`
component. Before, the `Entry`, while iterating sections, used the
index of that iteration to tell each Section whether it's the current
section and pass references to its adjacent sections.

This logic was moved into `Chapter`. The index of a section inside a
chapter cannot be used, though, to track the active scene of the
entry. Moreover, only references to adjacent sections inside that
chapter were passed.

Restore and move this logic into the `useEntryStructure` hook.

REDMINE-17300