Move paged-specific revision attributes to configuration
Migration strategy: For now, attribute columns remain, but are treated
as read-only, to be read as fallback in case the configuration column
has no value for the attribute key in question.

Since there shouldn't be values for those attributes in case of
scrolled entry type, we also migrate their column defaults to null.

REDMINE-17297