Extract entry template from theming
Account receives a 'first paged entry template', which stands in
for the default theming in places where attributes move from
theming to entry template. In a further iteration, we'll
introduce ways to configure accounts with multiple entry
templates, e.g. one for 'paged' and one for 'scrolled'.

As opposed to default theming, there is no belongs_to association
between account and first paged entry template. This means that
it needs special treatment with respect to CRUD.

In the admin backend, Formtastic's checkbox input, in principle,
allows setting true and false Boolean values as checked_value and
unchecked_value, but they are not persisted in a way that plays
nice e.g. with corresponding checkboxes in the editor.

For this reason we use a crutch of setting 'true' and 'false'
strings, as opposed to real Boolean values, as checked_value and
unchecked_value and converting them to Boolean values right in
the accounts admin, so that as soon as the values arrive at the
server, they are converted to the format that we also use in the
editor.

Also, for home_button_enabled, we ensure consistent data by doing a
replacement for '0' and '1' strings, which leads to the same data
format that we also had before moving the attribute to config. This
should soon vanish in a later PR, which will introduce something like
a schema for entry type-dependent config.

We rename home_button_enabled_by_default to home_button_enabled in the
entry template config, because this config is copied verbatim to new
entries' revision configurations. Revision configurations already
expect home_button_enabled as the config attribute's name.

REDMINE-17305