Filter new models in watchCollection resets
When returning to the outline while a content element is being
created, the `SectionThumbnail` included the new content element in
the collections, causing it to render without a key. This triggered a
React warning about unique keys in lists which made Capybara specs
fail in some cases.

Filtering them out made some specs fail, that used models without key
attributes.

To prevent the entry metadata which is watched inside an ad-hoc
collection from being filtered out, we need to make sure a key
attribute is present. We therefore add the possibility to provide
constant attributes via functions when watching collections.

REDMINE-17430