Use i18n-js from npm package in Webpack builds
Previously `i18n-js` was defined as external in the Webpack build, to
use `window.I18n` which is provided by the `i18n-js` gem via the asset
pipeline.

We want to use the `i18n-js` package for `pageflow-scrolled`, but
`window.I18n` is not defined in published scrolled entries since we do
not want to depend on the asset pipeline there. We therefore install
the `i18n-js` package via Yarn instead.

We still map the `i18n-js` import to globals in `iife` Rollup
outputs. That way nothing changes for the paged editor, paged frontend
and non editor code (i.e. `pageflow-analytics`) which uses
`pageflow/ui`.

In the scrolled editor, we now have two `I18n` objects: One on
`window` which is filled with translations via the asset pipeline, one
inside the bundle. During its setup, the `i18n-js` package looks for
an existing `I18n` object in the global scope to fall back
to. Unfortunately Webpack changes the global scope in a way which
makes it not point to window. Hence, inside the bundle, we use a
different `I18n` object. We therefore manually sync the two `I18n`
objects in `startEditor`.

For the scrolled frontend, we now have an empty `I18n` object
available, which we will fill in the next commit.

We can remove the vendored `i18n-js` file which was previously used in
Jest. The `useFakeTranslations` test helper did not reset translations
correctly in teardown, which caused problems with the new `i18n-js`
package.

REDMINE-17375