Create videos in seeds
Use transcoded videos in Backdrop and inlineVideo components.
Since in CI there is no transcoding, introduce special handling of
the storybook seed for videos:
When seeding the database for the specific "Storybook seed" entry,
video transcoding is omitted and the video file is set directly to "encoded".
Storybook can then render the video tag based on the existing video, even
though the video url points to a non existing file.