Add stories and seeds
- in seeds.rb, rewrite file references also for inlineBeforeAfter
content element, specifically its before_id and after_id props
- also in seeds.rb, remove obsolete starts_with? exception for
beforeAfter. This was anticipated for the case of having a
before/after as background, but we don't want that right
now. Before/After in foreground can be made full-width, and placing
additional content elements on a before/after sounds like a bad idea
to me, because it could lead to strange sliding behavior
- include church images in storybook rake task
- include storybook examples for all settings that make sense to
display in a storybook

REDMINE-17372