Allow setting type name on entry create
Contains parts of [1].

We add an entries/entry_types endpoint to serve the entry types that
are available for the currently selected account on an entry
form. They are used to show and hide entry type select options
accordingly.

The select input is only shown when there is more than one option
available.

We add two entries helpers. One is used to provide the initial
collection of all entry types to populate the select field, while the
other one is used to determine which entry types to hide or show.

Apart from JS-based client-side validation, there is also server-side
validation to ensure only available entry types could be set.

As a drive-by improvement, we look up an entry's entry type in the
global config instead of its own config. This reduces complexity and
could prevent e.g. display issues that might arise from config changes
on accounts and/or entries.

Type name of an entry can be found on show page.

Also, config_for(target) is made resilient against hypothetical future
targets that also have a `type_name` attribute but aren't entries.

[1] https://github.com/DanielKuhn/pageflow/commit/5526bffc2f6bea5629532a733d3c2a3bcafec223

REDMINE-17508