Fix help link in editor page type drop down
While replacing references to the global `pageflow.editor` with api
objects passed in via view options, it was overlooked that in the
event handler `this.options` does not refer to the options of the
`EditPageView`, but instead of the `ExtendedSelectInputView`.

Using an arrow function corrects the context of the callback.

REDMINE-17480