Prevent including pageflow/frontend in pageflow-paged/editor
The paged editor needs to import `Object` from `pageflow/frontend`. By
default, this causes Rollup to include the whole file in
`pageflow-paged/editor`. We therefore need to declare
`pageflow/frontend` as external. We can map the import to the
`pageflow` global, which provdes all exports of `pageflow/frontend`
for to be backwards compatible.

Doing the same for `pageflow-paged/frontend` lets us move `EntryData`
into `frontend` the module.