Resolve fadeVolume also when cancelled
Prevent warnings about unhandled promise rejection. We are either not
interested whether fading was cancelled (e.g. when using the volume
slider) or do not want to abort a flow just because volume changed
(e.g. changing volume during play and fade in). For fade out and pause
this was already handled explicitly. This case can now be simplified.