Rewrite build-edge script to use gemspec files
Before, the list of build outputs needed to be kept up to date in the
`gemspec` as well as in the `build-edge` script. Make `build-edge` use
file list from gemspec as single source of truth. This ensures that
things that have been tested on edge branches keep working in the
released gem.