Change default of text-shadow for gradient box
Instead of setting text-shadow default for all content inside
a gradient box, export withShadow class and use it only on
headlines inside gradient boxes for now.
This makes the CSS class available for all content elements
which desire a text shadow when placed inside a gradient box,
but removes the necessity to explicitly overwrite the text-shadow
if not.