Provide locale via separate context
The main motivation for this change is to allow `useI18n` outside of
an `EntryStateProvider` by giving the `LocaleContext` `en` as default
value. This lets us use translations in low level components without
having to resort to tools like `renderInEntry`.

As a nice side effect, this change prevents having all components with
translations subscribe to the entry state context. Below a
`React.memo`, components with `useI18n` will now only rerender when
the locale changes.

REDMINE-17673