Prevent running media loading spinner animation twice
Make sure the CSS class that triggers the CSS animation is not set
when the component is rendered on the server side. Once the component
is mounted, change state to apply the class.

This also ensures that the loading spinner will never be hidden before
`window.onload` since widgets are only mounted after that. Before,
choosing a small animation duration could reveal a still broken
looking entry.

REDMINE-17701