Hide focused controls if focus outline hidden
When controls (e.g. the player button) are clicked they receive the
focus. This causes the controls to never fade, even though the focus
outline is hidden (since it is only visible after keyboard
interaction). Only prevent controls from becoming inactive because the
focus is inside when focus rect is visible.

REDMINE-17751