Extend editor api to detect browser
Some parts of editor are not mobile or touch optimized,
so it makes sense to only support desktop browsers and
block editor on  mobile devices.

Editor api is extended to let entry types detect the type of browser and
block access if necessary, showing a block access view page.

We extend the api with a method, create a backbone view in respective entry type and finally
entry type calls the method and display the view if blocked or proceed
with the editor.

We extend api spec for the new method.

We also extend the creating_entry_type api documentation to include the
newly created method.