Make entry templates editable
We update Rubocop to allow checking block length of
'Pageflow::ActiveAdmin.register'. Before Rubocop 0.55.0, only block
names like 'register' were allowed.

Logic moves from accounts admin to a new entry templates admin.

Before this change, only one entry template for paged entry type was
editable through the account form, and a scrolled entry template would
be initialized on creation of the first scrolled entry, but not
persisted to the database.

We add an entry template admin for new/create and edit/update
actions. It is accessible only from an entry template tab on account
show page of accounts that have already been created.

We need to customize breadcrumbs and redirects, since there isn't an
index page or a show page for entry templates. To make it clear to the
user what they are currently editing, we customize the form's page
title, as well.

The account now has a list of existing and potential entry templates
that is used to populate the entry templates tab on its show
page. This'll probably change at some point when we allow multiple
entry templates per account/entry type.

Entry template provides translated entry type name as a virtual
attribute for use in several places in the admin.

Even though account publishers cannot access the account show page,
they would be able to edit entry templates using a link, if they have
it.

Entry template details move from the core of accounts' show page to
the entry template tab, since they are duplicated from now on.

The account form hints where the form fields have gone. There isn't a
link to the entry templates tab, because for account create, this link
wouldn't work yet.

For now, the entry templates form hard-codes the fact that currently
available config and widget inputs are for the paged entry type. When
we make the analytics widget available for scrolled entry type, this
is likely to change.

Entry templates tab contains entries for all existing entry templates
and all those that could be created according to enabled entry
types. If an entry type is disabled, its corresponding entry template
is not displayed here anymore. The tab also doubles as a kind of
show/index page for entry templates, containing data that previously
was per account because of the 1:1 mapping of accounts to themings.

REDMINE-17306